Move communicator
