+++ dbos/core.py
@@ -19,6 +19,7 @@ from dbos.context import (
     DBOSContextEnsure,
     DBOSContextSwap,
     EnterDBOSChildWorkflow,
+    EnterDBOSCommunicator,
     EnterDBOSTransaction,
     EnterDBOSWorkflow,
     OperationType,
@@ -28,6 +29,7 @@ from dbos.context import (
     get_local_dbos_context,
 )
 from dbos.error import (
+    DBOSCommunicatorMaxRetriesExceededError,
     DBOSNonExistentWorkflowError,
     DBOSRecoveryError,
     DBOSWorkflowConflictUUIDError,
@@ -42,7 +44,11 @@ from dbos.registrations import (
     set_dbos_func_name,
 )
 from dbos.roles import check_required_roles
-from dbos.system_database import WorkflowInputs, WorkflowStatusInternal
+from dbos.system_database import (
+    OperationResultInternal,
+    WorkflowInputs,
+    WorkflowStatusInternal,
+)
 
 if TYPE_CHECKING:
     from dbos.dbos import DBOS, Workflow, WorkflowHandle, WorkflowStatus
@@ -468,6 +474,117 @@ def _transaction(
                 ), "Transactions must be called from within workflows"
                 with DBOSAssumeRole(rr):
                     return invoke_tx(*args, **kwargs)
+            else:
+                tempwf = dbos.workflow_info_map.get("<temp>." + func.__qualname__)
+                assert tempwf
+                return tempwf(*args, **kwargs)
+
+        def temp_wf(*args: Any, **kwargs: Any) -> Any:
+            return wrapper(*args, **kwargs)
+
+        wrapped_wf = _workflow_wrapper(dbos, temp_wf)
+        set_dbos_func_name(temp_wf, "<temp>." + func.__qualname__)
+        dbos._register_wf_function(get_dbos_func_name(temp_wf), wrapped_wf)
+
+        return cast(F, wrapper)
+
+    return decorator
+
+
+def _communicator(
+    dbos: "DBOS",
+    *,
+    retries_allowed: bool = False,
+    interval_seconds: float = 1.0,
+    max_attempts: int = 3,
+    backoff_rate: float = 2.0,
+) -> Callable[[F], F]:
+    def decorator(func: F) -> F:
+
+        def invoke_comm(*args: Any, **kwargs: Any) -> Any:
+
+            attributes: TracedAttributes = {
+                "name": func.__name__,
+                "operationType": OperationType.COMMUNICATOR.value,
+            }
+            with EnterDBOSCommunicator(attributes) as ctx:
+                comm_output: OperationResultInternal = {
+                    "workflow_uuid": ctx.workflow_uuid,
+                    "function_id": ctx.function_id,
+                    "output": None,
+                    "error": None,
+                }
+                recorded_output = dbos.sys_db.check_operation_execution(
+                    ctx.workflow_uuid, ctx.function_id
+                )
+                if recorded_output:
+                    if recorded_output["error"] is not None:
+                        deserialized_error = utils.deserialize(recorded_output["error"])
+                        raise deserialized_error
+                    elif recorded_output["output"] is not None:
+                        return utils.deserialize(recorded_output["output"])
+                    else:
+                        raise Exception("Output and error are both None")
+                output = None
+                error = None
+                local_max_attempts = max_attempts if retries_allowed else 1
+                max_retry_interval_seconds: float = 3600  # 1 Hour
+                local_interval_seconds = interval_seconds
+                for attempt in range(1, local_max_attempts + 1):
+                    try:
+                        output = func(*args, **kwargs)
+                        comm_output["output"] = utils.serialize(output)
+                        error = None
+                        break
+                    except Exception as err:
+                        error = err
+                        if retries_allowed:
+                            dbos.logger.warning(
+                                f"Communicator being automatically retried. (attempt {attempt} of {local_max_attempts}). {traceback.format_exc()}"
+                            )
+                            ctx.get_current_span().add_event(
+                                f"Communicator attempt {attempt} failed",
+                                {
+                                    "error": str(error),
+                                    "retryIntervalSeconds": local_interval_seconds,
+                                },
+                            )
+                            if attempt == local_max_attempts:
+                                error = DBOSCommunicatorMaxRetriesExceededError()
+                            else:
+                                time.sleep(local_interval_seconds)
+                                local_interval_seconds = min(
+                                    local_interval_seconds * backoff_rate,
+                                    max_retry_interval_seconds,
+                                )
+
+                comm_output["error"] = (
+                    utils.serialize(error) if error is not None else None
+                )
+                dbos.sys_db.record_operation_result(comm_output)
+                if error is not None:
+                    raise error
+                return output
+
+        fi = get_or_create_func_info(func)
+
+        @wraps(func)
+        def wrapper(*args: Any, **kwargs: Any) -> Any:
+            rr: Optional[str] = check_required_roles(func, fi)
+            # Entering communicator is allowed:
+            #  In a communicator already, just call the original function directly.
+            #  In a workflow (that is not in a transaction / comm already)
+            #  Not in a workflow (we will start the single op workflow)
+            ctx = get_local_dbos_context()
+            if ctx and ctx.is_communicator():
+                # Call the original function directly
+                return func(*args, **kwargs)
+            if ctx and ctx.is_within_workflow():
+                assert (
+                    ctx.is_workflow()
+                ), "Communicators must be called from within workflows"
+                with DBOSAssumeRole(rr):
+                    return invoke_comm(*args, **kwargs)
             else:
                 tempwf = dbos.workflow_info_map.get("<temp>." + func.__qualname__)
                 assert tempwf
